Add concept of Multipurpose Sensor (as opposed to SimpleSensor)
In contrast to a Simple Sensor, which produces one kind of measurement
both in terms of the thing it is sensing and what it is sensing,
a MultipurposeSensor does more than one thing.

A Multipurpose Sensor can produce values for multiple telemetry names and multiple nodes.

For example, a PowerMeter can have two CTs. Let's say we have a heat pump
(a.heatpump) and a boost element (a.elt1), and we have a CT on the line to
each of these. Additionally, the PowerMeter provide multiple ElectricalQuantities
that we want to track. For now we have AmpsRms (Rms stands for root mean squared,
i.e. sqrt(2) times the max value of the sine wave of the AC current.) We will
shortly also have VoltsRms, PowerRms, and PhaseAngle. And there will be others.

In any case, values from the PowerMeter (a Multipurpose Sensor) need to specify:
  - _what_ physical object they are sensing (TelemetryTuple.AboutNode)
  - _what_ sensor was doing the sensing (TelemetryTuple.SensorNode)
  - _what_ the units and exponents of the reading are (TelemetryTuple.TelemetryName)