fix last commit. hw1 goes back to w
test_load_real_house should give way to a schema check on a house
model.

We need this now, because the house model changes per house and is
thus really a configuration setting, and we are starting to have
multiple real-life setups.

Note that these will go into `hybrid` universes. That means the
universe can combine simulated objects with objects in the real world.