Move atn <-> scada state management to slow dispatch contracts w heartbeats   - calculation of energy moves from synth generator to contract manager   - atn and scada both save latest heartbeat for contract   - concept of representation status: can set to Dormant in the .env file and then the scada will not accept inbound slow contracts
The major functional difference is that either or both the atn and scada can reboot or
lose connection and the contract continues. Indeed if the scada loses Internet it will maintain
its current contract through the end time.

The scada will stay in `Atn` mode for a grace period of 5 minutes after the latest contract
expires.

If the system goes into admin it will keep the existing contract going but will not start a new
one.